Found that there are differences between mac and linux db
Need to find out whether DocumentNotes table edits change the
gui on linux, or it's the same as Mac - the Documents table
notes field dominates. Does it matter? It does depending on what
is synced. Given that the Documents table is not the same on mac
and linux, try changing DocumentNotes and syncing.