Fixed bugs with export for mac
On mac, the notes are kept up to date in the Documents table (not
the DocumentNotes table, which seems defunct). It may transpire that
this is also the case on linux, but for now I have split the script
to detect the platform and work as before if using linux. It will
break on windows at the moment.

Also did a little bit of filename cleansing, and added an option to
ignore completely empty notes.